+++ src/pages/send/send.scss
@@ -1,8 +1,4 @@
 page-send {
-  .show-more {
-    display: flex;
-    text-align: center;
-  }
   .title {
     font-weight: 500;
     img {
@@ -18,22 +14,6 @@ page-send {
     justify-content: center;
     text-align: center;
     padding: 30px;
-  }
-  .wallet-warning {
-    ion-label {
-      white-space: inherit;
-    }
-  }
-  .wallet-disabled {
-    background-color: #fbfbfb;
-    ion-icon {
-      img {
-        background-color: #d4d4d4 !important;
-      }
-    }
-    .item-title {
-      color: #d4d4d4;
-    }
   }
   .send-header-wrapper {
     padding: 10px;
@@ -66,21 +46,6 @@ page-send {
     }
     .backspace-icon {
       z-index: 999;
-    }
-    .bitcoin-icon {
-      margin-right: 1rem;
-      margin-top: 1rem;
-      margin-bottom: 1rem;
-      width: 1rem;
-      height: 1rem;
-      // custom icons
-      &[class*='search-btc'] {
-        mask-image: url(../assets/img/icon-bitcoin-symbol.svg);
-        mask-size: 16px;
-        mask-position: 50% 50%;
-        mask-repeat: no-repeat;
-        background: #cccccc;
-      }
     }
   }
   .invalid {
