+++ src/pages/send/send.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { Events, NavController, NavParams } from 'ionic-angular';
 import * as _ from 'lodash';
@@ -6,20 +6,17 @@ import { Observable } from 'rxjs/Observable';
 
 // Providers
 import { ActionSheetProvider } from '../../providers/action-sheet/action-sheet';
-import { AddressBookProvider } from '../../providers/address-book/address-book';
 import { AddressProvider } from '../../providers/address/address';
 import { AppProvider } from '../../providers/app/app';
 import { ExternalLinkProvider } from '../../providers/external-link/external-link';
 import { IncomingDataProvider } from '../../providers/incoming-data/incoming-data';
 import { Logger } from '../../providers/logger/logger';
-import { PopupProvider } from '../../providers/popup/popup';
 import { ProfileProvider } from '../../providers/profile/profile';
-import { Coin, WalletProvider } from '../../providers/wallet/wallet';
+import { Coin } from '../../providers/wallet/wallet';
 import { WalletTabsProvider } from '../wallet-tabs/wallet-tabs.provider';
 
 // Pages
 import { WalletTabsChild } from '../wallet-tabs/wallet-tabs-child';
-import { AmountPage } from './amount/amount';
 import { MultiSendPage } from './multi-send/multi-send';
 
 export interface FlatWallet {
@@ -43,22 +40,10 @@ export class SendPage extends WalletTabsChild {
   public search: string = '';
   public walletsBtc;
   public walletsBch;
-  public walletBchList: FlatWallet[];
-  public walletBtcList: FlatWallet[];
-  public contactsList = [];
-  public filteredContactsList = [];
-  public filteredWallets = [];
   public hasBtcWallets: boolean;
   public hasBchWallets: boolean;
-  public hasContacts: boolean;
-  public contactsShowMore: boolean;
-  public amount: string;
-  public fiatAmount: number;
-  public fiatCode: string;
   public invalidAddress: boolean;
 
-  private CONTACTS_SHOW_LIMIT: number = 10;
-  private currentContactsPage: number = 0;
   private scannerOpened: boolean;
   private validDataTypeMap: string[] = [
     'BitcoinAddress',
@@ -71,11 +56,8 @@ export class SendPage extends WalletTabsChild {
     navCtrl: NavController,
     private navParams: NavParams,
     profileProvider: ProfileProvider,
-    private walletProvider: WalletProvider,
-    private addressBookProvider: AddressBookProvider,
     private logger: Logger,
     private incomingDataProvider: IncomingDataProvider,
-    private popupProvider: PopupProvider,
     private addressProvider: AddressProvider,
     private events: Events,
     walletTabsProvider: WalletTabsProvider,
@@ -86,6 +68,9 @@ export class SendPage extends WalletTabsChild {
   ) {
     super(navCtrl, profileProvider, walletTabsProvider);
   }
+
+  @ViewChild('transferTo')
+  transferTo;
 
   ionViewDidLoad() {
     this.logger.info('Loaded: SendPage');
@@ -101,84 +86,10 @@ export class SendPage extends WalletTabsChild {
     this.walletsBch = this.profileProvider.getWallets({ coin: 'bch' });
     this.hasBtcWallets = !_.isEmpty(this.walletsBtc);
     this.hasBchWallets = !_.isEmpty(this.walletsBch);
-    this.walletBchList = this.getBchWalletsList();
-    this.walletBtcList = this.getBtcWalletsList();
-    this.updateContactsList();
   }
 
   ionViewWillLeave() {
     this.events.unsubscribe('update:address');
-  }
-
-  private getBchWalletsList(): FlatWallet[] {
-    return this.hasBchWallets ? this.getRelevantWallets(this.walletsBch) : [];
-  }
-
-  private getBtcWalletsList(): FlatWallet[] {
-    return this.hasBtcWallets ? this.getRelevantWallets(this.walletsBtc) : [];
-  }
-
-  private getRelevantWallets(rawWallets): FlatWallet[] {
-    return rawWallets
-      .map(wallet => this.flattenWallet(wallet))
-      .filter(wallet => this.filterIrrelevantRecipients(wallet));
-  }
-
-  private updateContactsList(): void {
-    this.addressBookProvider.list().then(ab => {
-      this.hasContacts = _.isEmpty(ab) ? false : true;
-      if (!this.hasContacts) return;
-
-      let contactsList = [];
-      _.each(ab, (v, k: string) => {
-        contactsList.push({
-          name: _.isObject(v) ? v.name : v,
-          address: k,
-          network: this.addressProvider.validateAddress(k).network,
-          email: _.isObject(v) ? v.email : null,
-          recipientType: 'contact',
-          coin: this.addressProvider.validateAddress(k).coin,
-          getAddress: () => Promise.resolve(k)
-        });
-      });
-      this.contactsList = contactsList.filter(c =>
-        this.filterIrrelevantRecipients(c)
-      );
-      let shortContactsList = _.clone(
-        this.contactsList.slice(
-          0,
-          (this.currentContactsPage + 1) * this.CONTACTS_SHOW_LIMIT
-        )
-      );
-      this.filteredContactsList = _.clone(shortContactsList);
-      this.contactsShowMore =
-        this.contactsList.length > shortContactsList.length;
-    });
-  }
-
-  private flattenWallet(wallet): FlatWallet {
-    return {
-      color: wallet.color,
-      name: wallet.name,
-      recipientType: 'wallet',
-      coin: wallet.coin,
-      network: wallet.network,
-      m: wallet.credentials.m,
-      n: wallet.credentials.n,
-      isComplete: wallet.isComplete(),
-      needsBackup: wallet.needsBackup,
-      getAddress: () => this.walletProvider.getAddress(wallet, false)
-    };
-  }
-
-  private filterIrrelevantRecipients(recipient: {
-    coin: string;
-    network: string;
-  }): boolean {
-    return this.wallet
-      ? this.wallet.coin === recipient.coin &&
-          this.wallet.network === recipient.network
-      : true;
   }
 
   public shouldShowZeroState() {
@@ -196,11 +107,6 @@ export class SendPage extends WalletTabsChild {
     );
     await Observable.timer(250).toPromise();
     infoSheet.present();
-  }
-
-  public showMore(): void {
-    this.currentContactsPage++;
-    this.updateContactsList();
   }
 
   public openScanner(): void {
@@ -295,46 +201,47 @@ export class SendPage extends WalletTabsChild {
     });
   }
 
-  public processInput(): void {
-    if (this.search && this.search.trim() != '') {
-      this.searchWallets();
-      this.searchContacts();
+  public cleanSearch() {
+    this.search = '';
+    this.invalidAddress = false;
+  }
 
-      if (
-        this.filteredContactsList.length === 0 &&
-        this.filteredWallets.length === 0
+  public async processInput() {
+    if (this.search == '') this.invalidAddress = false;
+    const hasContacts = await this.checkIfContact();
+    if (!hasContacts) {
+      const parsedData = this.incomingDataProvider.parseData(this.search);
+      if (parsedData && parsedData.type == 'PayPro') {
+        const coin: string =
+          this.search.indexOf('bitcoincash') === 0 ? Coin.BCH : Coin.BTC;
+        this.incomingDataProvider
+          .getPayProDetails(this.search)
+          .then(payProDetails => {
+            payProDetails.coin = coin;
+            const isValid = this.checkCoinAndNetwork(payProDetails, true);
+            if (isValid) this.redir();
+          })
+          .catch(err => {
+            this.invalidAddress = true;
+            this.logger.warn(err);
+          });
+      } else if (
+        parsedData &&
+        _.indexOf(this.validDataTypeMap, parsedData.type) != -1
       ) {
-        const parsedData = this.incomingDataProvider.parseData(this.search);
-        if (parsedData && parsedData.type == 'PayPro') {
-          const coin: string =
-            this.search.indexOf('bitcoincash') === 0 ? Coin.BCH : Coin.BTC;
-          this.incomingDataProvider
-            .getPayProDetails(this.search)
-            .then(payProDetails => {
-              payProDetails.coin = coin;
-              const isValid = this.checkCoinAndNetwork(payProDetails, true);
-              if (isValid) this.redir();
-            })
-            .catch(err => {
-              this.invalidAddress = true;
-              this.logger.warn(err);
-            });
-        } else if (
-          parsedData &&
-          _.indexOf(this.validDataTypeMap, parsedData.type) != -1
-        ) {
-          const isValid = this.checkCoinAndNetwork(this.search);
-          if (isValid) this.redir();
-        } else {
-          this.invalidAddress = true;
-        }
+        const isValid = this.checkCoinAndNetwork(this.search);
+        if (isValid) this.redir();
       } else {
-        this.invalidAddress = false;
+        this.invalidAddress = true;
       }
     } else {
-      this.updateContactsList();
-      this.filteredWallets = [];
+      this.invalidAddress = false;
     }
+  }
+
+  public async checkIfContact() {
+    await Observable.timer(50).toPromise();
+    return this.transferTo.hasContactsOrWallets;
   }
 
   private checkIfLegacy(): boolean {
@@ -344,52 +251,6 @@ export class SendPage extends WalletTabsChild {
         this.search
       )
     );
-  }
-
-  public searchWallets(): void {
-    if (this.hasBchWallets && this.wallet.coin === 'bch') {
-      this.filteredWallets = this.walletBchList.filter(wallet => {
-        return _.includes(wallet.name.toLowerCase(), this.search.toLowerCase());
-      });
-    }
-    if (this.hasBtcWallets && this.wallet.coin === 'btc') {
-      this.filteredWallets = this.walletBtcList.filter(wallet => {
-        return _.includes(wallet.name.toLowerCase(), this.search.toLowerCase());
-      });
-    }
-  }
-
-  public searchContacts(): void {
-    this.filteredContactsList = _.filter(this.contactsList, item => {
-      let val = item.name;
-      return _.includes(val.toLowerCase(), this.search.toLowerCase());
-    });
-  }
-
-  public goToAmount(item): void {
-    item
-      .getAddress()
-      .then((addr: string) => {
-        if (!addr) {
-          // Error is already formated
-          this.popupProvider.ionicAlert('Error - no address');
-          return;
-        }
-        this.logger.debug('Got address:' + addr + ' | ' + item.name);
-        this.navCtrl.push(AmountPage, {
-          recipientType: item.recipientType,
-          amount: parseInt(this.navParams.data.amount, 10),
-          toAddress: addr,
-          name: item.name,
-          email: item.email,
-          color: item.color,
-          coin: item.coin,
-          network: item.network
-        });
-      })
-      .catch(err => {
-        this.logger.error('Send: could not getAddress', err);
-      });
   }
 
   public goToMultiSendPage(): void {
