+++ src/pages/send/transfer-to-modal/transfer-to-modal.scss
@@ -1,91 +1,2 @@
 page-transfer-to-modal {
-  .show-more {
-    display: flex;
-    text-align: center;
-  }
-  .title {
-    font-weight: 500;
-    img {
-      vertical-align: sub;
-      margin-right: 5px;
-    }
-  }
-  .no-wallet-message {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-    padding: 30px;
-  }
-  .wallet-warning {
-    ion-label {
-      white-space: inherit;
-    }
-  }
-  .wallet-disabled {
-    background-color: #fbfbfb;
-    ion-icon {
-      img {
-        background-color: #d4d4d4 !important;
-      }
-    }
-    .item-title {
-      color: #d4d4d4;
-    }
-  }
-  .send-header-wrapper {
-    padding: 10px;
-    background-color: white;
-    box-shadow: 0px 5px 10px 0px #cccccc;
-  }
-  .send-heading {
-    font-size: 15px;
-    padding: 0 0 16px 0;
-    border: none;
-    font-weight: 500;
-  }
-  .search-wrapper {
-    margin-top: 15px;
-    background-color: #f2f2f2;
-    display: flex;
-    align-items: baseline;
-    border: 1px solid #f2f2f2;
-    border-radius: 4px;
-    .item-inner {
-      align-items: center;
-      margin-left: 0rem !important;
-      border-bottom: 0px !important;
-      box-shadow: none !important;
-    }
-    .scan-icon {
-      margin: 0;
-      padding: 0.6rem 1rem 1rem 1rem;
-      z-index: 999;
-    }
-    .backspace-icon {
-      z-index: 999;
-    }
-    .bitcoin-icon {
-      margin-right: 1rem;
-      margin-top: 1rem;
-      margin-bottom: 1rem;
-      width: 1rem;
-      height: 1rem;
-      // custom icons
-      &[class*='search-btc'] {
-        mask-image: url(../assets/img/icon-bitcoin-symbol.svg);
-        mask-size: 16px;
-        mask-position: 50% 50%;
-        mask-repeat: no-repeat;
-        background: #cccccc;
-      }
-    }
-  }
-  .invalid {
-    border: 1px solid color($colors, error);
-    color: color($colors, error);
-    background-color: #fff6f7;
-  }
 }
